Fix application hanging on startup
This reverts one change from c4f1cec13e9116eded678c36f0033973dff45351
which had the side effect of `while (activeChannelIndex-- >= 0)` always
being true, resulting in an endless loop.

+++ Source/OutputChannelNames.h
@@ -65,7 +65,7 @@ private:
 public:
     juce::String getDeviceOutputChannelName(int activeChannelIndex);
     juce::StringArray getAllDeviceOutputChannelNames();
-    juce::String getInternalOutputChannelName(size_t activeChannelIndex);
+    juce::String getInternalOutputChannelName(int activeChannelIndex);
     void setInternalOutputChannelName(int activeChannelIndex, const juce::String& text);
 
 private:
